Rename and split tests into separate files
